requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
rich>=13.0.0
pyfiglet>=0.8.post1